--- src/Contacts/Contacts/App.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Configuration;
-using System.Data;
-using System.Windows;
-using View.Model.Services;
-using View.ViewModel;
-
-namespace Contacts
-{
-    /// <summary>
-    /// Interaction logic for App.xaml
-    /// </summary>
-    public partial class App : Application
-    {
-    }
-}
