+++ src/Contacts/Model/Model.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.11.35312.102
+VisualStudioVersion = 17.13.35818.85 d17.13
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "View", "Contacts\View.csproj", "{D76CC6D6-497E-47FC-BF66-A19F5FCA269B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Model", "Model.csproj", "{6DD5245B-E4CA-497E-B7AF-A26B6BDADF9C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D76CC6D6-497E-47FC-BF66-A19F5FCA269B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D76CC6D6-497E-47FC-BF66-A19F5FCA269B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D76CC6D6-497E-47FC-BF66-A19F5FCA269B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D76CC6D6-497E-47FC-BF66-A19F5FCA269B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6DD5245B-E4CA-497E-B7AF-A26B6BDADF9C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6DD5245B-E4CA-497E-B7AF-A26B6BDADF9C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6DD5245B-E4CA-497E-B7AF-A26B6BDADF9C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6DD5245B-E4CA-497E-B7AF-A26B6BDADF9C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C6B77444-4662-4752-AEAE-2FAC5490E20A}
+		SolutionGuid = {83372138-DA04-41B2-BD3C-56A66B46340A}
 	EndGlobalSection
 EndGlobal
