+++ src/Contacts/View/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="Contacts.App"
+﻿<Application x:Class="View.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:Contacts"
+             xmlns:local="clr-namespace:View"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
