+++ src/Contacts/View/AssemblyInfo.cs
@@ -1,6 +1,6 @@
 using System.Windows;
 
-[assembly: ThemeInfo(
+[assembly:ThemeInfo(
     ResourceDictionaryLocation.None,            //where theme specific resource dictionaries are located
                                                 //(used if a resource is not found in the page,
                                                 // or application resource dictionaries)
