+++ src/Contacts/View/Control/ContactControl.xaml.cs
