+++ src/Contacts/View/Converter/BoolToVisibilityConverter.cs
