+++ src/Contacts/View/View.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.13.35818.85 d17.13
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "View", "View.csproj", "{3B94841D-6F6B-4256-AD6C-9A4F56C9E888}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ViewModel", "..\ViewModel\ViewModel.csproj", "{A69CDC74-A8D1-82BE-8A86-332F0EDB6A3C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Model", "..\Model\Model.csproj", "{19C7C86B-88DA-9A98-B57D-7FA23AB4DA69}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3B94841D-6F6B-4256-AD6C-9A4F56C9E888}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3B94841D-6F6B-4256-AD6C-9A4F56C9E888}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3B94841D-6F6B-4256-AD6C-9A4F56C9E888}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3B94841D-6F6B-4256-AD6C-9A4F56C9E888}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A69CDC74-A8D1-82BE-8A86-332F0EDB6A3C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A69CDC74-A8D1-82BE-8A86-332F0EDB6A3C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A69CDC74-A8D1-82BE-8A86-332F0EDB6A3C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A69CDC74-A8D1-82BE-8A86-332F0EDB6A3C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{19C7C86B-88DA-9A98-B57D-7FA23AB4DA69}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{19C7C86B-88DA-9A98-B57D-7FA23AB4DA69}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{19C7C86B-88DA-9A98-B57D-7FA23AB4DA69}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{19C7C86B-88DA-9A98-B57D-7FA23AB4DA69}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B493EFD9-D2B1-404A-9946-D103B6D2C92E}
+	EndGlobalSection
+EndGlobal
