+++ src/Contacts/ViewModel/ViewModel.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.13.35818.85
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ViewModel", "ViewModel.csproj", "{C6F2E644-8520-4D11-BDB4-1BC9EEC42C7B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Model", "..\Model\Model.csproj", "{19C7C86B-88DA-9A98-B57D-7FA23AB4DA69}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C6F2E644-8520-4D11-BDB4-1BC9EEC42C7B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6F2E644-8520-4D11-BDB4-1BC9EEC42C7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6F2E644-8520-4D11-BDB4-1BC9EEC42C7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6F2E644-8520-4D11-BDB4-1BC9EEC42C7B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{19C7C86B-88DA-9A98-B57D-7FA23AB4DA69}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{19C7C86B-88DA-9A98-B57D-7FA23AB4DA69}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{19C7C86B-88DA-9A98-B57D-7FA23AB4DA69}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{19C7C86B-88DA-9A98-B57D-7FA23AB4DA69}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {040A8274-68F9-4511-AB11-C50A755789A4}
+	EndGlobalSection
+EndGlobal
